refactor: move test out of integration tests
This test doesn't need to be integration, on the contrary, it can
benefit from within-crate visibility.
